requirements.txt
Django==4.1.3
django-environ==0.9.0
django-extensions==3.2.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
factory-boy==3.2.1
Pillow==9.3.0
pytest-django==4.5.2
psycopg2-binary==2.9.5
whitenoise==6.2.0
